Moves anomalyzer to nupic.data.generators module
