+++ nupic/data/generators/anomalyzer.py
